Rename `base.py -> vlm_base.py`
