+++ vllm/model_executor/models/vlm_base.py
